LibG: Small improvements

+++ libs/libui/Window.cpp
@@ -46,7 +46,7 @@ void Window::receive_event(UniquePtr<LFoundation::Event> event)
     if (event->type() == Event::Type::MouseActionEvent) {
         if (m_superview) {
             MouseActionEvent& own_event = *(MouseActionEvent*)event.get();
-            auto* view = m_superview->hit_test({own_event.x(), own_event.y()});
+            auto* view = m_superview->hit_test({(int)own_event.x(), (int)own_event.y()});
             view->receive_mouse_action_event(own_event);
         }
     }
